Also rename the pool
